update example implementation;

+++ sessions/recap-project-6/example-implementation/pages/places/[id]/index.js
@@ -1,10 +1,9 @@
-import Link from "next/link";
-import { useRouter } from "next/router.js";
+import { useRouter } from "next/router";
 import useSWR from "swr";
 import styled from "styled-components";
-import { StyledLink } from "../../../components/StyledLink.js";
-import { StyledButton } from "../../../components/StyledButton.js";
-import { StyledImage } from "../../../components/StyledImage.js";
+import { StyledLink } from "../../../components/StyledLink";
+import { StyledButton } from "../../../components/StyledButton";
+import { StyledImage } from "../../../components/StyledImage";
 
 const ImageContainer = styled.div`
   position: relative;
@@ -13,8 +12,9 @@ const ImageContainer = styled.div`
 
 const ButtonContainer = styled.section`
   display: flex;
+  flex-direction: column;
   justify-content: space-between;
-  gap: 0.2rem;
+  gap: 0.5rem;
 
   & > * {
     flex-grow: 1;
@@ -24,8 +24,9 @@ const ButtonContainer = styled.section`
 
 const StyledLocationLink = styled(StyledLink)`
   text-align: center;
-  background-color: white;
-  border: 3px solid lightsalmon;
+  background-color: lightgray;
+  color: black;
+  border: none;
 `;
 
 export default function DetailsPage() {
@@ -38,17 +39,20 @@ export default function DetailsPage() {
   if (!isReady || isLoading || error) return <h2>Loading...</h2>;
 
   async function deletePlace() {
-    await fetch(`/api/places/${id}`, {
-      method: "DELETE",
-    });
-    router.push("/");
+    if (confirm("are you sure you want to delete this place?")) {
+      await fetch(`/api/places/${id}`, {
+        method: "DELETE",
+      });
+      router.push("/");
+      return;
+    }
   }
 
   return (
     <>
-      <Link href={"/"} passHref legacyBehavior>
-        <StyledLink $justifySelf="start">back</StyledLink>
-      </Link>
+      <StyledLink href={"/"} $justifySelf="start">
+        back
+      </StyledLink>
       <ImageContainer>
         <StyledImage
           src={place.image}
